pyproject.toml
[tool.poetry]
name = "coredns-manager"
version = "0.1.0"
description = "CoreDNS Management Tool"
authors = ["Development Team <dev@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.122.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
sqlmodel = "^0.0.27"
python-dotenv = "^1.2.1"
jinja2 = "^3.1.6"
requests = "^2.32.5"
docker = "^7.1.0"
aiosqlite = "^0.21.0"
pydantic-settings = "^2.12.0"
itsdangerous = "^2.2.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"
black = "^25.11.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
